Styled components have been moved to separate files
