+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "./UI/Header";
+import Header from "./UI/Header/Header";
 
 const App = () => {
   return (
