+++ src/UI/Header/StyledHeader.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const StyledHeader = styled.header`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-top: 10px;
+  padding-bottom: 10px;
+`;
+
+export default StyledHeader;
