+++ src/components/Button/Button.js
@@ -1,0 +1,8 @@
+import React from "react";
+import StyledButton from "./StyledButton";
+
+const Button = ({ type, children }) => {
+  return <StyledButton type={type}>{children}</StyledButton>;
+};
+
+export default Button;
